refactor(kornia-imgproc): hoist check_size and sealed to kernel_common
