+++ crates/kornia-imgproc/src/color/hsv/mod.rs
@@ -5,25 +5,7 @@ mod kernels;
 
 // ===== Sealed-trait dispatch =========================================================
 
-mod sealed {
-    pub trait Sealed {}
-}
-
-#[inline]
-fn check_size<T, U, const C1: usize, const C2: usize, A1: ImageAllocator, A2: ImageAllocator>(
-    src: &Image<T, C1, A1>,
-    dst: &Image<U, C2, A2>,
-) -> Result<(), ImageError> {
-    if src.size() != dst.size() {
-        return Err(ImageError::InvalidImageSize(
-            src.cols(),
-            src.rows(),
-            dst.cols(),
-            dst.rows(),
-        ));
-    }
-    Ok(())
-}
+use crate::color::kernel_common::{check_size, sealed};
 
 /// Compile-time dispatch to the right RGB→HSV kernel for each pixel type.
 ///
@@ -45,7 +27,6 @@ pub trait RgbFromHsv: sealed::Sealed + Sized {
     ) -> Result<(), ImageError>;
 }
 
-impl sealed::Sealed for f32 {}
 impl HsvFromRgb for f32 {
     fn hsv_from_rgb_impl<A1: ImageAllocator, A2: ImageAllocator>(
         src: &Image<f32, 3, A1>,
@@ -67,7 +48,6 @@ impl RgbFromHsv for f32 {
     }
 }
 
-impl sealed::Sealed for f64 {}
 impl HsvFromRgb for f64 {
     fn hsv_from_rgb_impl<A1: ImageAllocator, A2: ImageAllocator>(
         src: &Image<f64, 3, A1>,
