+++ crates/kornia-imgproc/src/color/yuv/mod.rs
@@ -7,28 +7,7 @@ pub use kernels::{Packed422, Planar420};
 
 // ===== Family A: RGB ↔ YCbCr / YUV (planar 3-channel) ===============================
 
-mod sealed {
-    pub trait Sealed {}
-    impl Sealed for u8 {}
-    impl Sealed for f32 {}
-    impl Sealed for f64 {}
-}
-
-#[inline]
-fn check_size<T, U, const C1: usize, const C2: usize, A1: ImageAllocator, A2: ImageAllocator>(
-    src: &Image<T, C1, A1>,
-    dst: &Image<U, C2, A2>,
-) -> Result<(), ImageError> {
-    if src.size() != dst.size() {
-        return Err(ImageError::InvalidImageSize(
-            src.cols(),
-            src.rows(),
-            dst.cols(),
-            dst.rows(),
-        ));
-    }
-    Ok(())
-}
+use crate::color::kernel_common::{check_size, sealed};
 
 /// Compile-time dispatch to the right RGB↔{YCbCr,YUV} kernel for each pixel type.
 ///
